# GYMaster-API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GYMaster",
    "version": "v1"
  },
  "paths": {
...
    "/api/Usuarios/Validar": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CredencialesDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/Catalogo": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioGimnasios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UsuariosDTO"
              }
            }
          }
        }
      }
    },
    "/api/Usuarios/Buscar": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioUsuarios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UsuariosDTO"
            }
          }
        }
      }
    },
    "/api/Usuarios/Registrar": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuariosDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/Modificar": {
      "put": {
        "tags": [
          "Usuarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioUsuarios",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuariosDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/Actualizar": {
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CredencialesDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/Afiliar": {
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioUsuarios",
            "type": "string"
          },
          {
            "in": "query",
            "name": "UsuarioGimnasios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuarios/RegistroTarjeta": {
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioUsuarios",
            "type": "string"
          },
          {
            "in": "query",
            "name": "CodigoNFC",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HttpStatusCode"
            }
          }
        }
      }
    },
...
    "/api/Verificacion/Verificar": {
      "get": {
        "tags": [
          "Verificacion"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Token",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Verificacion/Registrar": {
      "post": {
        "tags": [
          "Verificacion"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/VerificacionDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "CredencialesDTO": {
      "type": "object",
      "properties": {
        "usuario": {
          "type": "string"
        },
        "contrasena": {
          "type": "string"
        }
      }
    },
...
    "HttpStatusCode": {
      "format": "int32",
      "enum": [
        100,
        101,
        102,
        103,
        200,
        201,
        202,
        203,
        204,
        205,
        206,
        207,
        208,
        226,
        300,
        301,
        302,
        303,
        304,
        305,
        306,
        307,
        308,
        400,
        401,
        402,
        403,
        404,
        405,
        406,
        407,
        408,
        409,
        410,
        411,
        412,
        413,
        414,
        415,
        416,
        417,
        421,
        422,
        423,
        424,
        426,
        428,
        429,
        431,
        451,
        500,
        501,
        502,
        503,
        504,
        505,
        506,
        507,
        508,
        510,
        511
      ],
      "type": "integer"
    },
...
    "UsuariosDTO": {
      "type": "object",
      "properties": {
        "nombres": {
          "type": "string"
        },
        "apellidos": {
          "type": "string"
        },
        "fechaNacimiento": {
          "format": "date-time",
          "type": "string"
        },
        "sexo": {
          "type": "string"
        },
        "estatura": {
          "format": "int32",
          "type": "integer"
        },
        "peso": {
          "format": "int32",
          "type": "integer"
        },
        "telefono": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "usuario1": {
          "type": "string"
        },
        "contrasena": {
          "type": "string"
        },
        "usuarioGimnasios": {
          "type": "string"
        },
        "fechaRegistro": {
          "format": "date-time",
          "type": "string"
        },
        "codigoTarjeta": {
          "type": "string"
        }
      }
    },
...
    "VerificacionDTO": {
      "type": "object",
      "properties": {
        "usuario": {
          "type": "string"
        },
        "permiso": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    }
  }
}